Code refactors + documentation improvements
- `src/Prelude/Prelude.ts` ---> `src/Lib/index.ts`
- `src/Prelude/Runtime/*` ---> `src/Lib/*`
