+++ src/Lib/index.ts
@@ -1,0 +1,16 @@
+/**
+ * @example
+ * ```typescript
+ * import * as Prelude from 'prelude'
+ * Prelude.eqInteger.eq(1n,2n)
+ * ```
+ *
+ * @module prelude
+ */
+
+export * from "./Prelude.js";
+export * from "./Maybe.js";
+export * from "./Either.js";
+export * from "./Json.js";
+export * from "./Ord.js";
+export * from "./Eq.js";
