--- src/Prelude/Prelude.ts
@@ -1,6 +1,0 @@
-export * from "./Runtime/Prelude.js";
-export * from "./Runtime/Maybe.js";
-export * from "./Runtime/Either.js";
-export * from "./Runtime/Json.js";
-export * from "./Runtime/Ord.js";
-export * from "./Runtime/Eq.js";
