+++ src/Tests/tsconfig.json
@@ -2,6 +2,6 @@
   "extends": "../tsconfig-base",
   "include": ["**/*"],
   "references": [
-    { "path": "../Prelude" }
+    { "path": "../Lib" }
   ]
 }
